move approximation to passes
(cherry picked from commit adc30334607808fa94d0c1d66f292a0ba073766b)
