+++ tests/passes/writeset_underapproximation_test.py
@@ -1,7 +1,7 @@
 # Copyright 2019-2023 ETH Zurich and the DaCe authors. All rights reserved.
 
 import dace
-from dace.sdfg.writeset_underapproximation import UnderapproximateWrites
+from dace.sdfg.analysis.writeset_underapproximation import UnderapproximateWrites
 from dace.subsets import Range
 from dace.transformation.pass_pipeline import Pipeline
